Add files via upload
Today there are many online video streaming services available some are regional, some are national and some of them serve all over the world like Netflix which has a huge customer base and are used by millions of people in the world. Netflix offers its users with access to internet shows in demand. It allows customers to take advantages of the internet and watch unlimited TV shows and movies. Moreover the company has a separate library of movies to which customers can subscribe too and watch instantly on their electronic devices, it also has its own productions known as Netflix Originals. Netflix is 2nd largest entertainment company by market capitalization as of today and 8th most trusted brand, It is available worldwide except in a handful of countries.